+++ NEWS.md
@@ -1,3 +1,13 @@
+# **wehoop 0.2.0**
+
+- Added support for ESPN's play-by-play endpoints with the addition of the following functions:
+- ```wehoop::wnba_espn_game_all()``` - a convenience wrapper function around the following three functions (returns the results as a list of three data frames)
+- ```wehoop::wnba_espn_team_box()```
+- ```wehoop::wnba_espn_player_box()```
+- ```wehoop::wnba_espn_pbp()```
+- ```wehoop::wnba_espn_teams()``` 
+- ```wehoop::wbb_espn_scoreboard()``` 
+
 # **wehoop 0.1.0**
 
 - Added support for ESPN's play-by-play endpoints with the addition of the following functions: